Triggering event when a token is activated
Adding a new event that will be triggered when a token is activated. This change is helpful for the case when you want to repond to such events, for example to scroll to the token in question.

I also built the project and include the dist files. It has some previous changes.